task_1_3/requirements.txt
google-genai
openai
python-dotenv
flask
flask-cors
